🎨 UI: #10 Level 3. View 파일 분리 1. HarryPotterSeriseView.swift 파일 이름 변경 → HomeView.swift    - 더 용도에 맞는 이름으로 변경 2. Level 2의 책 정보 영역을 InfoView.swift로 분리시킨다. 3. Level 3의 Dedication, Summary 영역 → LableContentView.swift로 분리시킨다.    - 각각의 구성을 해당 UIView를 재사용하여 UI를 구성한다.
